transaction-service/requirements.txt
kafka-python
faker